General Improvements
Improvements and additions. Added a Remove vehicle function and a remove Customer function but both are buggy. There is a chance the remove vehicle method works perfectly fine but the remove customer isn't. When the customer is deleted, then the user needs to run the program again, therefore the information is run again and the customer isn't actually deleted. A back/logout button might be introduced to solve this bug.